// packages/routes/src/types.d.ts
export type HttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH' | 'OPTIONS';
export interface HttpRequest {
    method: HttpMethod;
    url: string;
    path: string;
    query: Record<string, string | string[] | undefined>;
    params: Record<string, string>;
    headers: Record<string, string | string[] | undefined>;
    body?: unknown;
    files?: File[];
}
export interface HttpResponse {
    status: number;
    headers: Record<string, string>;
    body: unknown;
}